+++ CHANGELOG.rst
@@ -1,6 +1,9 @@
 Unreleased
 ----------
 
+Added
+ * Command-line interface
+
 Changed
  * Refreshing and updating the state of the devices is now synchronized to the
    reply from the server